Extract SkillReadFileTool into own submodule, improve proptest diagnostics
Replace `unwrap()` with `expect("failed to decode bytes as UTF-8")` in
the `utf8_boundary_size_succeeds` proptest so failure output includes
the decoding failure reason rather than a bare panic message.

Move `SkillReadFileTool` into `src/tools/builtin/skill_tools/read_file.rs`,
mirroring the existing `install.rs` and `remove.rs` submodule pattern.
Keeps the parent module under the repository file-size ceiling and groups
related code in a discoverable layout.
